created a transcriptTemplate and am now using it for shows episodes and npr episodes
