+++ pages/npr/[slug]/index.vue
